Output comment create/update body from Action
It turns out this GitHub Action can't comment on PRs that were opened by
a fork, because GitHub [only gives forked repos read access][1] to the
main repo. And GitHub [won't pass any other secrets][2] to forks, so you
can't pass in a custom personal access token that way, either.

The only solution I've found is making the actual comment create/update
request in a separate workflow that runs [on a `workflow_run` event][3]
and retrieves artifacts saved by the triggering workflow. This second
workflow will then have the write access required to comment on the PR
without having the ability to run untrusted code from forks at the same
time.

To allow this situation, this Action will need to output the data
required to make the API request, and I think outputting the whole
request body is the most straightforward way to do so.

[1]: https://docs.github.com/en/actions/security-guides/automatic-token-authentication#permissions-for-the-github_token
[2]: https://docs.github.com/en/actions/security-guides/encrypted-secrets#using-encrypted-secrets-in-a-workflow
[3]: https://securitylab.github.com/research/github-actions-preventing-pwn-requests/